contract-compliance-agent/requirements.txt
# Contract Compliance Agent - Dependencies
# Python 3.10+ required

# Core dependencies
httpx>=0.25.0              # Async HTTP client for Ollama API
pydantic>=2.0.0            # Data validation and schemas
pyyaml>=6.0               # YAML configuration parsing
rich>=13.0.0              # Beautiful terminal output
typer>=0.9.0              # CLI framework

# Document processing
pypdf>=3.0.0              # PDF text extraction
python-docx>=1.0.0        # DOCX text extraction

# Testing
pytest>=7.4.0             # Test framework
pytest-cov>=4.1.0         # Coverage reporting
pytest-asyncio>=0.21.0    # Async test support

# Development (optional)
black>=23.0.0             # Code formatting
ruff>=0.1.0               # Linting
mypy>=1.5.0               # Type checking
